Added unit test, reusing RegisterDriver
